+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor inventory state management to use useReducer for UI state and add new actions for toggling quick add and archived item visibility
 - Refactor useInventory and inventoryReducer to manage lastFetchedAt state with useReducer
 - Refactor FilterBarForm to replace useState with useRef for search term management and implement debouncing directly in the search handler
+- Refactor MainContainer to streamline useShoppingList integration
 
 ---